added user model folder and required that in the index folder. Also have a comment in server.js for the user
