+++ server/server.js
@@ -5,6 +5,7 @@ const path = require('path');
 
 const { typeDefs, resolvers } = require('./schemas');
 const db = require('./config/connection');
+// const { User } = require('./models');
 
 const PORT = process.env.PORT || 3001;
 const app = express();
